@@ -1,43 +1,85 @@
 {
-  "name": "addressbook--server-side",
-  "version": "0.1.0",
+  "name": "contacts-book",
+  "version": "1.0.0-alpha",
   "private": true,
-  "scripts": {
-    "ssr-build": "cd ./views && npm run ssr-build",
-    "start": "cd ./views && npm run build-dev && npm run build-ssr && nodemon ./bin/www",
-    "deploy": "npm run ssr-build && node ./bin/www"
+  "engines": {
+    "node": ">=10.15.3"
   },
+  "scripts": {
+    "test:watch": "jest --passWithNoTests --watch",
+    "test": "npm run test:watch",
+    "heroku-postbuild": "npm --prefix ./views install ./views",
+    "prepare-dev-frontend": "cd ./views && npm run build:dev",
+    "prepare-prod-frontend": "cd ./views && npm run build",
+    "start-dev-server": "DEBUG=contactsbook:* nodemon ./bin/www",
+    "start": "npm run start-dev-server",
+    "predeploy": "npm run prepare-prod-frontend",
+    "deploy": "NODE_ENV=production node ./bin/www",
+    "deploy-heroku": "NODE_ENV=production node ./bin/www",
+    "deploy-heroku-with-pm2": "NODE_ENV=production pm2-runtime start ecosystem.config.js --env production node ./bin/www"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/lvnam96/contactsbook.git"
+  },
+  "license": "MIT",
   "dependencies": {
-    "async": "^2.6.1",
+    "async": "^3.1.0",
     "bcryptjs": "^2.4.3",
-    "body-parser": "~1.18.2",
-    "connect-pg-simple": "^4.2.1",
-    "cookie-parser": "~1.4.3",
-    "debug": "^3.1.0",
-    "express": "^4.16.3",
-    "express-session": "^1.15.6",
-    "morgan": "~1.9.0",
-    "mysql": "^2.15.0",
+    "body-parser": "^1.19.0",
+    "compression": "^1.7.4",
+    "connect-pg-simple": "^6.0.1",
+    "cookie-parser": "^1.4.4",
+    "cors": "^2.8.5",
+    "debug": "^4.1.1",
+    "express": "^4.16.4",
+    "express-jwt": "^5.3.1",
+    "express-session": "^1.17.0",
+    "helmet": "^3.21.2",
+    "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.15",
+    "morgan": "^1.9.1",
     "node-gyp": "^3.6.2",
     "passport": "^0.4.0",
+    "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
-    "pg": "^7.4.3",
-    "pg-promise": "^8.4.6",
+    "pg": "^7.14.0",
+    "pg-promise": "^10.3.1",
+    "pm2": "^4.2.0",
     "pug": "^2.0.3",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
-    "serve-favicon": "^2.5.0"
+    "randomcolor": "^0.5.4",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "serve-favicon": "^2.5.0",
+    "supports-color": "^7.1.0"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.1",
-    "eslint": "^5.3.0",
-    "eslint-plugin-babel": "^5.2.1",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.1.1",
-    "eslint-plugin-node": "^8.0.0",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-react": "^7.11.0",
-    "nodemon": "^1.17.4",
+    "@babel/core": "^7.7.4",
+    "@typescript-eslint/eslint-plugin": "^2.10.0",
+    "@typescript-eslint/parser": "^2.10.0",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^24.9.0",
+    "eslint": "^6.7.2",
+    "eslint-config-prettier": "^6.7.0",
+    "eslint-config-standard": "^14.1.0",
+    "eslint-config-standard-jsx": "^8.1.0",
+    "eslint-config-standard-react": "^9.2.0",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-node": "^10.0.0",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-promise": "^4.1.1",
+    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-react-hooks": "^2.3.0",
+    "eslint-plugin-standard": "^4.0.1",
+    "jest": "^24.9.0",
+    "nodemon": "^2.0.1",
+    "prettier": "^1.19.1",
+    "prettier-eslint": "^9.0.1",
+    "react-axe": "^3.3.0",
+    "superagent": "^5.1.2",
+    "supertest": "^4.0.2",
     "webpack-node-externals": "^1.7.2"
   }
 }
