@@ -1,94 +1,149 @@
 {
-  "name": "adrsbook-front-end-serverside-rendering",
-  "version": "0.1.0",
-  "description": "adrsbook app's front-end using serverside rendering",
+  "name": "contactsbook--front-end",
+  "version": "1.0.0-b",
+  "private": true,
+  "description": "Front-end source code of a contacts book app",
   "main": "index.js",
+  "engines": {
+    "node": ">=10.15.3"
+  },
   "scripts": {
-    "webpack-ssr": "webpack --config webpack.config.ssr.js",
-    "build-ssr": "NODE_ENV=production mkdir -p ssr && npm run webpack-ssr",
-    "build-prod": "NODE_ENV=production webpack --config webpack.config.prod.js",
-    "build-dev": "webpack -p --config webpack.config.dev.js",
-    "prebuild": "npm run build-ssr",
-    "build": "npm run build-prod",
-    "start": "npm run mock-api | webpack-dev-server --config webpack.config.dev.js",
+    "build:ssr": "webpack --config webpack.config.ssr.js",
+    "build:app-dev": "NOVE_ENV=production webpack --config webpack.config.prod.js",
+    "build:dev": "FE_ENV=development run-s build:ssr build:app-dev",
+    "build:app-prod": "webpack --config webpack.config.prod.js",
+    "build:prod": "run-s build:ssr build:app-prod",
+    "build": "NODE_ENV=production FE_ENV=production npm run build:prod",
     "generate-random-data": "node ./mockAPI/generate.js",
-    "premock-api": "npm run generate-random-data",
-    "mock-api": "json-server --watch ./mockAPI/db.json --port 2004 --routes ./mockAPI/routes.json",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start:api-server": "json-server --watch ./mockAPI/db.json --port 2004 --delay 100 --routes ./mockAPI/routes.json --quiet",
+    "start:dev-api": "IS_DEV_API=true run-p generate-random-data start:api-server",
+    "start:api": "run-s generate-random-data start:api-server",
+    "start:dev-server": "webpack-dev-server --config webpack.config.dev.js",
+    "start": "run-p start:api start:dev-server",
+    "start:dev-backend": "npm run start:dev-server",
+    "test": "jest --passWithNoTests"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/lvnam96/addressbook--server-side.git"
+    "url": "git+https://github.com/lvnam96/contactsbook.git"
   },
   "author": "garyle",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/lvnam96/addressbook--server-side/issues"
+    "url": "https://github.com/lvnam96/contactsbook/issues"
   },
-  "homepage": "https://github.com/lvnam96/addressbook--server-side#readme",
+  "homepage": "https://github.com/lvnam96/contactsbook#readme",
   "devDependencies": {
-    "autoprefixer": "^8.6.5",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-jest": "^23.6.0",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "chance": "^1.0.16",
-    "clean-webpack-plugin": "^0.1.19",
-    "css-loader": "^0.28.11",
-    "cssnano": "^3.10.0",
+    "@babel/cli": "^7.7.4",
+    "@babel/core": "^7.7.4",
+    "@babel/plugin-proposal-class-properties": "^7.7.4",
+    "@babel/plugin-proposal-decorators": "^7.7.4",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.7.4",
+    "@babel/plugin-proposal-object-rest-spread": "^7.7.4",
+    "@babel/plugin-proposal-optional-chaining": "^7.7.4",
+    "@babel/plugin-syntax-dynamic-import": "^7.7.4",
+    "@babel/plugin-transform-react-jsx-source": "^7.7.4",
+    "@babel/plugin-transform-runtime": "^7.7.4",
+    "@babel/preset-env": "^7.7.4",
+    "@babel/preset-react": "^7.7.4",
+    "@babel/preset-typescript": "^7.7.4",
+    "@hot-loader/react-dom": "^16.11.0",
+    "@types/classnames": "^2.2.9",
+    "@types/lodash": "^4.14.149",
+    "@types/react": "^16.9.17",
+    "@types/react-axe": "^3.1.0",
+    "@types/react-dom": "^16.9.4",
+    "@types/react-redux": "^7.1.6",
+    "@types/react-router-dom": "^5.1.3",
+    "@types/webpack-env": "^1.14.1",
+    "@types/yup": "^0.26.27",
+    "acorn": "^7.1.0",
+    "autoprefixer": "^9.7.2",
+    "babel-loader": "^8.0.6",
+    "chance": "^1.1.3",
+    "chokidar": "^3.3.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^5.0.5",
+    "css-loader": "^3.2.0",
+    "cssnano": "^4.1.10",
     "faker": "^4.1.0",
-    "file-loader": "^1.1.11",
+    "file-loader": "^4.3.0",
     "html-webpack-plugin": "^3.2.0",
-    "image-webpack-loader": "^4.3.1",
-    "jest": "^23.6.0",
-    "json-schema-faker": "^0.5.0-rc15",
-    "json-server": "^0.14.0",
-    "mini-css-extract-plugin": "^0.4.2",
-    "node-sass": "^4.9.3",
-    "optimize-css-assets-webpack-plugin": "^4.0.3",
-    "postcss": "^6.0.23",
-    "postcss-import": "^11.1.0",
-    "postcss-loader": "^2.1.6",
-    "postcss-preset-env": "^5.3.0",
-    "postcss-scss": "^1.0.6",
-    "prop-types": "^15.6.2",
-    "pug": "^2.0.3",
+    "image-webpack-loader": "^6.0.0",
+    "json-schema-faker": "^0.5.0-rc23",
+    "json-server": "^0.15.1",
+    "media-query-plugin": "^1.3.1",
+    "mini-css-extract-plugin": "^0.8.0",
+    "node-sass": "^4.13.0",
+    "npm-run-all": "^4.1.5",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "postcss": "^7.0.23",
+    "postcss-import": "^12.0.1",
+    "postcss-loader": "^3.0.0",
+    "postcss-preset-env": "^6.7.0",
+    "postcss-scss": "^2.0.0",
+    "pug": "^2.0.4",
     "pug-loader": "^2.4.0",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
-    "react-hot-loader": "^4.3.6",
-    "regenerator-runtime": "^0.12.1",
-    "resolve-url-loader": "^2.3.1",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.21.0",
-    "url-loader": "^1.1.1",
-    "webpack": "^4.18.0",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.8",
-    "webpack-node-externals": "^1.7.2"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "resolve-url-loader": "^3.1.1",
+    "sass-loader": "^8.0.0",
+    "source-map-loader": "^0.2.4",
+    "style-loader": "^1.0.0",
+    "stylelint": "^12.0.0",
+    "stylelint-config-prettier": "^6.0.0",
+    "stylelint-config-recommended-scss": "^4.1.0",
+    "stylelint-prettier": "^1.1.1",
+    "stylelint-scss": "^3.13.0",
+    "typescript": "^3.7.2",
+    "url-loader": "^2.3.0",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0",
+    "webpack-node-externals": "^1.7.2",
+    "webpackbar": "^4.0.0"
   },
   "peerDependencies": {
-    "react": "^16.5.0",
-    "react-dom": "^16.5.0"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.3.1",
-    "axios": "^0.18.0",
-    "babel-polyfill": "^6.26.0",
-    "bootstrap": "^4.1.3",
-    "formik": "^1.3.1",
+    "@babel/runtime-corejs3": "^7.7.4",
+    "@fortawesome/fontawesome-free": "^5.11.2",
+    "@sentry/browser": "^5.9.1",
+    "axios": "^0.19.0",
+    "bootstrap": "^4.3.1",
+    "classnames": "^2.2.6",
+    "connected-react-router": "^6.6.0",
+    "core-js": "^3.4.2",
+    "formik": "^2.0.6",
+    "history": "^4.10.1",
+    "js-cookie": "^2.2.1",
     "jssha": "^2.3.1",
-    "lodash": "^4.17.10",
-    "luxon": "^1.4.3",
+    "lodash": "^4.17.15",
+    "luxon": "^1.21.2",
+    "md5": "^2.2.1",
+    "memoize-one": "^5.1.1",
     "normalize.css": "^8.0.0",
-    "react-modal": "^3.6.1",
-    "react-redux": "^5.0.7",
-    "react-virtualized": "^9.20.1",
-    "redux": "^4.0.0",
+    "popper.js": "^1.16.0",
+    "prop-types": "^15.7.2",
+    "randomcolor": "^0.5.4",
+    "react-axe": "^3.3.0",
+    "react-bootstrap": "^1.0.0-beta.16",
+    "react-color": "^2.17.3",
+    "react-hot-loader": "^4.12.18",
+    "react-modal": "^3.11.1",
+    "react-redux": "^7.1.3",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
+    "react-select": "^3.0.8",
+    "react-virtualized": "^9.21.2",
+    "react-virtualized-auto-sizer": "^1.0.2",
+    "react-window": "^1.8.5",
+    "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
-    "yup": "^0.26.6"
+    "regenerator-runtime": "^0.13.3",
+    "uuid": "^3.3.3",
+    "yup": "^0.27.0"
   }
 }
