@@ -1,9 +1,11 @@
-const configObj = {
-    user: 'garyle',
-    host: 'localhost',
-    database: 'adrsbook',
-    password: 'bdd8OC0qgd',
-    port: 5432
-};
+let configObj = require('./db-config.json');
+
+if (process.env.DATABASE_URL) {
+  // for heroku
+  configObj = {
+    connectionString: process.env.DATABASE_URL,
+    ssl: true,
+  };
+}
 
 module.exports = configObj;
