@@ -4,8 +4,13 @@ const isIterable = (obj) => {
     return false;
   }
   return typeof obj[Symbol.iterator] === 'function';
-}
+};
+
+const checkValidUUID = (str) =>
+  typeof str === 'string' &&
+  str.match(/^[0-9a-f]{8}-[0-9a-f]{4}-[0-5][0-9a-f]{3}-[089ab][0-9a-f]{3}-[0-9a-f]{12}$/i) !== null;
 
 module.exports = {
-    isIterable
+  isIterable,
+  checkValidUUID,
 };
