@@ -1,27 +1,64 @@
+/* eslint-disable camelcase */
 // map data from account table in database
 // to obj with right props which class User needs
-const mapUserData  = (userDataFromDB) => {
-    const { id, username, password, email, facebook_id, birth, phone, nicename, created_on, last_login, salt } = userDataFromDB;
-    return {
-        id,
-        uname: username,
-        username,
-        passwd: password,
-        password,
-        fbid: facebook_id,
-        facebookId: facebook_id,
-        email,
-        nicename,
-        birth,
-        phone,
-        created_on,
-        createdOn: created_on,
-        last_login,
-        lastLogin: last_login,
-        salt
-    };
+const mapUserData = ({
+  id,
+  username,
+  password,
+  email,
+  facebook_id,
+  birth,
+  phone,
+  nicename,
+  created_on,
+  salt,
+  cbooks,
+  last_activated_cbook_id,
+  last_login,
+  is_active,
+}) => {
+  // const {
+  //   id,
+  //   username,
+  //   password,
+  //   email,
+  //   facebook_id,
+  //   birth,
+  //   phone,
+  //   nicename,
+  //   created_on,
+  //   salt,
+  //   cbooks,
+  //   last_activated_cbook_id,
+  //   last_login,
+  //   is_active,
+  // } = userDataFromDB;
+  return {
+    id,
+    uname: username,
+    passwd: password,
+    fbId: facebook_id,
+    email,
+    nicename,
+    birth,
+    phone,
+    createdOn: created_on,
+    salt,
+    meta: {
+      lastActivatedCbookId: last_activated_cbook_id,
+      lastLogin: last_login,
+      isActive: is_active,
+    },
+    cbooks,
+  };
+};
+
+const formatIdsList = (ids, isResultString = false) => {
+  const formattedArr = ids.map((id) => "'" + id + "'");
+  return isResultString ? formattedArr.join(',') : formattedArr;
 };
 
 module.exports = {
-    mapUserData
+  mapUserData,
+  formatIdsList,
 };
