@@ -1,8 +1,8 @@
-const dbUtils = require('./dbUtils'),
-    random = require('./random'),
-    checker = require('./checker');
+const dbUtils = require('./dbUtils');
+const random = require('./random');
+const checker = require('./checker');
 module.exports = {
-    dbUtils,
-    random,
-    ...checker
+  dbUtils,
+  random,
+  ...checker,
 };
