@@ -2,41 +2,70 @@ const express = require('express');
 const router = express.Router();
 const User = require('../../classes/User');
 const auth = require('../../services/auth');
-const db = require('../../db/index');
-const adrsbkRouter = require('./addressbook');
+const cbookRouter = require('./cbook');
 const contactsRouter = require('./contacts');
+const userRouter = require('./user');
 const adbk = require('../../classes/adbk');
+const { extractJWT } = require('../../middlewares/jwt');
+const { extractAndValidateXsrfToken } = require('../../middlewares/csrf');
 
 // '/backdoor' route
 
+const middlewares = [extractJWT, auth.allowUserAccessing, extractAndValidateXsrfToken];
+
+// this route need to be secured (using token) & limited on the amount of requests
 router.get('/is-uname-used', (req, res, next) => {
-    User.isUnameUsed(req.query.uname, (customErrMsg, result) => {
-        if (customErrMsg) {
-            console.log('Error on server');
-        }
-        res.json({ res: result });
+  User.findByUname(req.query.uname)
+    .then((user) => {
+      res.json({ res: !!user });
+    })
+    .catch((err) => {
+      console.error(err);
+      next(err);
     });
 });
 
-router.get('/get-all-data', (req, res, next) => {
-    adbk.user.loadAll(req.user.id).then(data => {
-        data = {
-            user: data.user.toJSON(),
-            adrsbook: data.adrsbook.toJSON()
-        };
-        res.json({ res: true, data });
-    }).catch(err => {
-        console.error(err);
-        res.json({ res: false });
+router.get('/is-email-used', (req, res, next) => {
+  User.findByEmail(req.query.email)
+    .then((user) => {
+      res.json({ res: !!user });
+    })
+    .catch((err) => {
+      console.error(err);
+      next(err);
     });
 });
 
-router.use('/addressbook', auth.restrictNonUserMiddleware, adrsbkRouter);
+router.get('/get-all-data', ...middlewares, (req, res, next) => {
+  adbk.user
+    .loadAll(req.user.id)
+    .then((resData) => {
+      const json = {
+        res: true,
+        data: {
+          ...resData,
+          user: resData.user.toJSON(),
+          contacts: resData.contacts.toJSON(),
+        },
+      };
+      res.json(json);
+      return json;
+    })
+    .catch((err) => {
+      console.error(err);
+      res.json({ res: false });
+    });
+});
 
-router.use('/contacts', auth.restrictNonUserMiddleware, contactsRouter);
+router.use('/cbook', ...middlewares, cbookRouter);
 
-router.get('/', auth.restrictNonUserMiddleware, (req, res, next) => {
-    console.log('GET request to /backdoor');
+router.use('/contacts', ...middlewares, contactsRouter);
+
+router.use('/user', ...middlewares, userRouter);
+
+router.get('/', ...middlewares, (req, res, next) => {
+  res.sendStatus(403);
+  res.end();
 });
 
 module.exports = router;
