@@ -3,8 +3,9 @@ const router = express.Router();
 
 // '/users' route
 
-router.get('/', function(req, res, next) {
-	res.send('respond with a resource');
+router.get('/', (req, res, next) => {
+  res.sendStatus(404);
+  next();
 });
 
 module.exports = router;
