@@ -1,7 +1,10 @@
-// import 'babel-polyfill';
+import 'core-js/stable';
+import 'regenerator-runtime/runtime';
 // import React from 'react';
 // import ReactDOM from 'react-dom';
 
 import '../scss/styles.scss';
+
+import './shims/requestIdleCallback';
 
 // document.addEventListener('DOMContentLoaded', () => {});
