@@ -1,18 +1,29 @@
-import SignInFormContainer from './js/SignInFormContainer';
+import 'react-hot-loader';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './js/App.jsx';
 // import registerServiceWorker from './registerServiceWorker';
 
 // import './scss/styles.scss';
 
+const hydrate = (App) => {
+  ReactDOM.hydrate(<App />, document.getElementsByClassName('signin-form-wrapper')[0]);
+};
+
 document.addEventListener('DOMContentLoaded', () => {
-
-    ReactDOM.hydrate(
-        <SignInFormContainer />,
-        document.getElementsByClassName('signin-form-wrapper')[0]
-    );
-
+  hydrate(App);
 });
 
-if (process.env.NODE_ENV !== 'production') {
-    module.hot.accept();
+if (core.status.isDev) {
+  if (module.hot) {
+    module.hot.accept('./js/App.jsx', () => {
+      const HotReloadedApp = require('./js/App.jsx').default;
+      hydrate(HotReloadedApp);
+    });
+  }
+
+  var axe = require('react-axe');
+  axe(React, ReactDOM, 1000);
 }
+
 // registerServiceWorker();
